Add "quiet" mode for use with cron.
In some organizations wsgetmail is run very frequently out of cron for
many different configuration files.  If the output of wsgetmail isn't
discarded that will generate a torrent of emails.  If stdout is discared
then emails will only be generated on error, but the error messages
don't identify the configuration that's in use.

This commit adds a "quiet" mode that suppresses the normal messages on
stdout, but preserves them so they can be replayed when there are any
errors in order to identify the configuration file (i.e. account or
folder) that had the error.